requirements.txt
pydantic==2.6.1
uvicorn==0.27.1
fastapi==0.109.2
google-cloud-firestore==2.14.0
google-cloud-pubsub==2.19.4
python-dotenv==1.0.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
google-cloud-logging==3.9.0
gunicorn==21.2.0